# cmd/relay/openapi.yaml
openapi: 3.1.0
info:
  title: Relay Webhook Delivery API
  description: |
    Relay is a webhook delivery system that guarantees delivery through
    idempotency keys, persistent storage, exponential backoff retries,
    and circuit breakers.

    ## Authentication
    All API requests require an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    API requests are rate limited per API key. Check response headers for limits:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets
  version: 1.0.0
  contact:
    name: Relay Support
    url: https://github.com/relay/relay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.relay.example.com
    description: Production
...
paths:
...
  /api/v1/connectors:
    get:
      tags:
        - Connectors
      summary: List connectors
      description: Returns all registered connectors
      operationId: listConnectors
      responses:
        "200":
          description: List of connectors
          content:
            application/json:
              schema:
                type: object
                properties:
                  connectors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Connector"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Connectors
      summary: Create a connector
      description: Creates a new connector for notifications or integrations
      operationId: createConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectorRequest"
      responses:
        "201":
          description: Connector created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Connector with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/connectors/{connectorName}:
    get:
      tags:
        - Connectors
      summary: Get connector by name
      description: Returns a single connector by its name
      operationId: getConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorName"
      responses:
        "200":
          description: Connector details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Connectors
      summary: Update a connector
      description: Updates an existing connector configuration
      operationId: updateConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnectorRequest"
      responses:
        "200":
          description: Connector updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Connectors
      summary: Delete a connector
      description: Deletes a connector
      operationId: deleteConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorName"
      responses:
        "204":
          description: Connector deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # Metrics Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
...
    ConnectorName:
      name: connectorName
      in: path
      required: true
      description: Connector name
      schema:
        type: string
  schemas:
...
    Connector:
      type: object
      properties:
        name:
          type: string
          description: Unique name for the connector
        type:
          $ref: "#/components/schemas/ConnectorType"
        enabled:
          type: boolean
          description: Whether the connector is active
        config:
          $ref: "#/components/schemas/ConnectorConfig"
        templates:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ConnectorTemplate"
          description: Message templates keyed by notification type
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - type
        - enabled
        - config
        - createdAt
        - updatedAt
    ConnectorType:
      type: string
      enum:
        - slack
        - email
        - webhook
        - pagerduty
      description: Type of integration connector
    ConnectorConfig:
      type: object
      properties:
        webhookUrl:
          type: string
          format: uri
          description: Webhook URL (for slack, webhook types)
        smtpHost:
          type: string
          description: SMTP server host (for email type)
        smtpPort:
          type: integer
          description: SMTP server port (for email type)
        smtpUser:
          type: string
          description: SMTP username (for email type)
        smtpFrom:
          type: string
          format: email
          description: From address for emails (for email type)
        apiKey:
          type: string
          description: API key (for pagerduty type)
        routingKey:
          type: string
          description: Routing/service key (for pagerduty type)
    ConnectorTemplate:
      type: object
      properties:
        subject:
          type: string
          description: Subject line template (for email)
        body:
          type: string
          description: Message body template
        format:
          type: string
          enum:
            - text
            - html
            - markdown
          description: Format of the message body
    CreateConnectorRequest:
      type: object
      properties:
        name:
          type: string
          description: Unique name for the connector
          minLength: 1
          maxLength: 255
        type:
          $ref: "#/components/schemas/ConnectorType"
        enabled:
          type: boolean
          default: true
        config:
          $ref: "#/components/schemas/ConnectorConfig"
        templates:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ConnectorTemplate"
      required:
        - name
        - type
        - config
    UpdateConnectorRequest:
      type: object
      properties:
        enabled:
          type: boolean
        config:
          $ref: "#/components/schemas/ConnectorConfig"
        templates:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ConnectorTemplate"

    # Metrics Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
      required:
        - error
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid request body"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid API key"
            code: "UNAUTHORIZED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Event not found"
            code: "NOT_FOUND"
...
security:
  - ApiKeyAuth: []